ref: Refactor to integration
