+++ packages/vue/src/index.ts
@@ -1,6 +1,7 @@
 export * from '@sentry/browser';
 
-export { init, initVueApp } from './sdk';
+export { init } from './sdk';
 export { vueRouterInstrumentation } from './router';
 export { attachErrorHandler } from './errorhandler';
 export { createTracingMixins } from './tracing';
+export { VueIntegration } from './integration';
