+++ packages/vue/src/types.ts
@@ -25,16 +25,14 @@ export type ViewModel = {
   };
 };
 
-export interface Options extends TracingOptions, BrowserOptions {
+export interface VueOptions extends TracingOptions {
   /** Vue constructor to be used inside the integration (as imported by `import Vue from 'vue'` in Vue2) */
   Vue?: Vue;
 
   /**
-   * Vue app instance(s) to be used inside the integration (as generated by `createApp` in Vue3 )
-   * Set this to `false` to indicate you are purposefully _not_ setting up a Vue app right now,
-   * e.g. if you want to manually call `initVueApp` later.
+   * Vue app instance(s) to be used inside the integration (as generated by `createApp` in Vue3).
    */
-  app?: Vue | Vue[] | false;
+  app?: Vue | Vue[];
 
   /**
    * When set to `false`, Sentry will suppress reporting of all props data
@@ -51,6 +49,8 @@ export interface Options extends TracingOptions, BrowserOptions {
   /** {@link TracingOptions} */
   tracingOptions?: Partial<TracingOptions>;
 }
+
+export interface Options extends BrowserOptions, VueOptions {}
 
 /** Vue specific configuration for Tracing Integration  */
 export interface TracingOptions {
