+++ packages/vue/test/integration/init.test.ts
@@ -1,6 +1,10 @@
 import { createApp } from 'vue';
 
+import { VueIntegration } from '../../src/integration';
+import type { Options } from '../../src/types';
 import * as Sentry from './../../src';
+
+const PUBLIC_DSN = 'https://username@domain/123';
 
 describe('Sentry.init', () => {
   let warnings: unknown[] = [];
@@ -22,9 +26,8 @@ describe('Sentry.init', () => {
       template: '<div>hello</div>',
     });
 
-    Sentry.init({
+    runInit({
       app,
-      defaultIntegrations: false,
     });
 
     app.mount(el);
@@ -38,10 +41,9 @@ describe('Sentry.init', () => {
       template: '<div>hello</div>',
     });
 
-    Sentry.init({
+    runInit({
       // this is a bit "hacky", but good enough to test what we want
       Vue: app,
-      defaultIntegrations: false,
     });
 
     app.mount(el);
@@ -57,9 +59,8 @@ describe('Sentry.init', () => {
 
     app.mount(el);
 
-    Sentry.init({
+    runInit({
       app,
-      defaultIntegrations: false,
     });
 
     expect(warnings).toEqual([
@@ -73,9 +74,7 @@ describe('Sentry.init', () => {
       template: '<div>hello</div>',
     });
 
-    Sentry.init({
-      defaultIntegrations: false,
-    });
+    runInit({});
 
     app.mount(el);
 
@@ -86,15 +85,16 @@ Update your \`Sentry.init\` call with an appropriate config option:
     ]);
   });
 
-  it('does not warn when passing app=false', () => {
+  it('does not warn when skipping Vue integration', () => {
     const el = document.createElement('div');
     const app = createApp({
       template: '<div>hello</div>',
     });
 
     Sentry.init({
-      app: false,
+      dsn: PUBLIC_DSN,
       defaultIntegrations: false,
+      integrations: [],
     });
 
     app.mount(el);
@@ -102,3 +102,23 @@ Update your \`Sentry.init\` call with an appropriate config option:
     expect(warnings).toEqual([]);
   });
 });
+
+function runInit(options: Partial<Options>): void {
+  const hasRunBefore = Sentry.getCurrentHub().getIntegration(VueIntegration);
+
+  const integration = new VueIntegration();
+
+  Sentry.init({
+    dsn: PUBLIC_DSN,
+    defaultIntegrations: false,
+    integrations: [integration],
+    ...options,
+  });
+
+  // Because our integrations API is terrible to test, we need to make sure to check
+  // If we've already had this integration registered before
+  // if that's the case, `setup()` will not be run, so we need to manually run it :(
+  if (hasRunBefore) {
+    integration['_setupIntegration'](Sentry.getCurrentHub());
+  }
+}
